refactor: Split tests and fixtures
